Clear the compiler's warning list down to two
The backlog named five deprecations; a forced recompile of every module found
twenty-four warnings, and all but two of them close mechanically.

`currentWindowAdaptiveInfo()` becomes `currentWindowAdaptiveInfoV2()`. Nothing
downstream moves: both questions `AppScaffold` asks are breakpoint comparisons
against the real width, so a 1400dp window already answered yes to each before
it had an L class of its own.

`LocalLifecycleOwner` now comes from `androidx.lifecycle.compose` at all three
sites, which costs a catalog entry and one `implementation` line in `:core:camera`
and `:feature:profile`. The alias left behind in `androidx.compose.ui.platform`
still forwards there, so the alternative was three warnings kept forever to
avoid two dependency lines on an artifact already in the graph.

Thirteen redundant casts were `listSaver` restores casting values whose type the
saver already infers. `ExportTest`'s three `payload.profile!!` become one local —
Kotlin only flagged the second and third, since the first is what smart-casts the
rest. Two unnecessary safe calls, one unused `Unit`, and a test whose name carried
a `"` that Windows cannot put in a report filename.

What is left needs a decision rather than a call swap, and the backlog entry now
says so: the swipe-to-dismiss `confirmValueChange`, whose replacement is dynamic
anchors and whose `DECISIONS.md` entry argues the delete happens inside the
callback; and `announceForAccessibility`, which Compose 1.12 answers only with
`liveRegion`.

Verified: `assembleDebug` and `testDebugUnitTest` green, and the only warnings
left are those two.

+++ feature/progress/src/main/java/ph/mart/healthapp/feature/progress/ui/pressure/LogBloodPressureState.kt
@@ -22,9 +22,9 @@ internal class LogBloodPressureState(form: BloodPressureForm = BloodPressureForm
             restore = { saved ->
                 LogBloodPressureState(
                     form = BloodPressureForm(
-                        systolic = saved[0] as Int,
-                        diastolic = saved[1] as Int,
-                        pulseBpm = saved[2] as Int,
+                        systolic = saved[0],
+                        diastolic = saved[1],
+                        pulseBpm = saved[2],
                     ),
                 )
             },
